+++ harbinger.py
@@ -1,5 +1,6 @@
 from configparser import ConfigParser
 from datetime import datetime
+from os import listdir
 
 import discord
 from discord.ext import commands
@@ -20,19 +21,11 @@ class Harbinger:
     server_public_ip = Configure.reveal(config["Server"]["server_public_ip"])
     moderator_role_id = int(Configure.reveal(config["Roles"]["moderator"]))
     developer_role_id = int(Configure.reveal(config["Roles"]["developer"]))
+    d_time = int(config["Bot"]["delete_after"])
     rgb = config["Custom Color"]["rgb"]
     r, g, b = map(int, rgb.split())
     custom_color = discord.Color.from_rgb(int(r), int(g), int(b))
-    cogs = (
-        "cogs.moderation",
-        "cogs.status",
-        "cogs.help",
-        "cogs.tools",
-        "cogs.minecraft",
-        "cogs.music",
-        "cogs.8ball",
-        "cogs.notes",
-    )
+
     start_time = datetime.now()
 
     intents = discord.Intents.default()
@@ -47,10 +40,14 @@ class Harbinger:
     @bot.event
     async def setup_hook() -> None:
         """Sequentially load cogs."""
-        for cog in cogs:
-            print(f"Loading {cog}...")
-            await bot.load_extension(cog)
-            print(f"Loaded {cog}.")
+        print(f"Loading cogs...")
+        for cog in listdir("cogs"):
+            if cog.endswith(".py") == True:
+                try:
+                    await bot.load_extension(f"cogs.{cog[:-3]}")
+                    print(f"Loaded {cog}")
+                except Exception as exc:
+                    print(f"An error has occured: {exc}.")
 
     def get_ver() -> str:
         """Check CHANGELOG.md for version info, return version string.
@@ -83,7 +80,6 @@ class Harbinger:
 
 
 bot = Harbinger.bot
-cogs = Harbinger.cogs
 
 
 def main():